% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{dict}
\alias{dict}
\title{Function to match comments and colors}
\usage{
dict(x, pattern, replace)
}
\arguments{
\item{x}{The comments to change}

\item{pattern}{the value to change}

\item{replace}{the value which should be replaced}
}
\description{
Function to match comments and colors
}

