% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{extract_timeline}
\alias{extract_timeline}
\title{Extract calling timeline}
\usage{
extract_timeline(interact, func, time = 1, calledBy = "NA")
}
\arguments{
\item{interact}{the interaction matrix create by interact function}

\item{func}{function of interest for which the timeline should be created}

\item{time}{the iterative time}

\item{calledBy}{the function which calls func}
}
\description{
Extract calling timeline
}

