% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{find_clusters}
\alias{find_clusters}
\title{Finds clusters of interactions}
\usage{
find_clusters(GraphLab)
}
\arguments{
\item{GraphLab}{output from GraphLab function}
}
\description{
Finds clusters of interactions
}

