% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{get_comments}
\alias{get_comments}
\title{Extract comments from functions}
\usage{
get_comments(filename)
}
\arguments{
\item{filename}{: filenames from which we should extract comments}
}
\description{
Extract comments from functions
}

