% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{GraphLab}
\alias{GraphLab}
\title{GraphLab}
\usage{
GraphLab(path = "")
}
\arguments{
\item{path}{: path to the R/ folder with scripts}
}
\description{
Statuses for the gTag can be one of complete, ongoing, undoc (umented), unknown
}
\author{
Paul Deveau <paul.deveau at curie.fr>

Pierre Gestraud
}

