% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{PlotGraphLab}
\alias{PlotGraphLab}
\title{PlotGraphLab}
\usage{
PlotGraphLab(GraphLab, func, filterOut = c("base", "utils"))
}
\arguments{
\item{GraphLab}{Output of the GraphLab function for the whole folder}

\item{func}{The function of interest for which the interaction graph should be plotted}

\item{filterOut}{name of packages from which the functions should be ignored. By default: base & utils}
}
\description{
Creates an interaction plot based on the output of GraphLab for a given function
}
\examples{
G<-GraphLab(system.file("extdata", "", package = "RDevGraphlab"))
# Same as DevGraphLab in this example:
PlotGraphLab(GraphLab = G, func = "Start")
# Graph for \\code{progeny} function only
PlotGraphLab(GraphLab = G,func = "progeny")
# Ignoring ggplot2 imports:
PlotGraphLab(GraphLab = G,func = "Start", filterOut = "ggplot2")
}

