% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{DevGraphLab}
\alias{DevGraphLab}
\title{DevGraphLab}
\usage{
DevGraphLab(path, filterOut = c("base", "utils"))
}
\arguments{
\item{path}{Path to the folder with all R scripts for the package}
}
\description{
Plots graph of interaction for all functions in the package
}

