% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GraphLab.R
\name{PlotGraphLab}
\alias{PlotGraphLab}
\title{PlotGraphLab}
\usage{
PlotGraphLab(GraphLab, func, filterOut = c("base", "utils"))
}
\arguments{
\item{func}{The function of interest for which the interaction graph should be plotted}

\item{filterOut}{name of packages from which the functions should be ignored. By default: base & utils}

\item{Graphlab}{Output of the GraphLab function for the whole folder}
}
\description{
Creates an interaction plot based on the output of Graphlab for a given function
}

